refactor/employee: team error messaging updated

+++ src/api-gateway/src/team/team.service.ts
@@ -1,17 +1,13 @@
-import {
-  ConflictException,
-  forwardRef,
-  Inject,
-  Injectable,
-} from '@nestjs/common';
+import { forwardRef, Inject, Injectable } from '@nestjs/common';
 import { CreateTeamDto } from './dto/create-team.dto';
 import { UpdateTeamDto } from './dto/update-team.dto';
-import { Types, FlattenMaps } from 'mongoose';
+import { Types } from 'mongoose';
 import { Team } from './entities/team.entity';
 import { EmployeeService } from '../employee/employee.service';
 import { CompanyService } from '../company/company.service';
 import { JobService } from '../job/job.service';
 import { TeamRepository } from './team.repository';
+import { ValidationResult } from 'src/auth/entities/validationResult.entity';
 
 @Injectable()
 export class TeamService {
@@ -28,7 +24,7 @@ export class TeamService {
   async validateCreateTeam(team: CreateTeamDto) {
     // Check if the company exists
     if (!(await this.companyService.companyIdExists(team.companyId))) {
-      throw new ConflictException('Company not found');
+      return new ValidationResult(false, `Company not found`);
     }
 
     // Check if the team members exist in the company
@@ -39,9 +35,7 @@ export class TeamService {
           team.companyId,
         ))
       ) {
-        throw new ConflictException(
-          `Team member ${memberId.toString()} not found`,
-        );
+        return new ValidationResult(false, `Team member not found`);
       }
     }
 
@@ -53,7 +47,7 @@ export class TeamService {
           team.companyId,
         ))
       ) {
-        throw new ConflictException('Team leader not found');
+        return new ValidationResult(false, `Team leader not found`);
       }
     }
 
@@ -61,70 +55,65 @@ export class TeamService {
     if (
       (await this.findByNameInCompany(team.teamName, team.companyId)) != null
     ) {
-      throw new ConflictException('Team already exists');
+      return new ValidationResult(false, `Team already exists`);
     }
   }
 
   async validateUpdateTeam(teamId: Types.ObjectId, team: UpdateTeamDto) {
+    // Check if the team exists
+    if (!(await this.teamExists(teamId))) {
+      return new ValidationResult(false, `Team not found`);
+    }
+
     // Getting the company ID from the team
     const teamTemp = await this.findById(teamId);
     const companyId = teamTemp.companyId;
-    console.log('companyId: ', companyId);
-    console.log('one');
+
     // Check if the team exists in the company
     if ('teamMembers' in team && team.teamMembers) {
-      console.log('A');
       for (const memberId of team.teamMembers) {
-        console.log('member: ', memberId);
         if (
           !(await this.employeeService.employeeExistsForCompany(
             memberId,
             companyId,
           ))
         ) {
-          throw new ConflictException(
-            `Team member ${memberId.toString()} not found`,
-          );
+          return new ValidationResult(false, `Team member not found`);
         }
       }
     }
-    console.log('two');
     // Check if the team leader exists in the company
     if ('teamLeaderId' in team && team.teamLeaderId) {
       if (!(await this.employeeService.employeeExists(team.teamLeaderId))) {
-        throw new ConflictException('Team leader not found');
+        return new ValidationResult(false, `Team leader not found`);
       }
     }
-    console.log('three');
     // Check if the jobs exist in the company
     if ('currentJobAssignments' in team && team.currentJobAssignments) {
       for (const jobId of team.currentJobAssignments) {
         if (!(await this.jobService.jobExistsInCompany(jobId, companyId))) {
-          throw new ConflictException(
-            `Job assignment ${jobId.toString()} not found`,
-          );
+          return new ValidationResult(false, `Job assignment not found`);
         }
       }
     }
-    console.log('four');
     // Check if the team already exists in the company
-
     if ((await this.findByNameInCompany(team.teamName, companyId)) != null) {
-      throw new ConflictException('Team already exists');
+      return new ValidationResult(false, `Team already exists`);
     }
   }
 
   async create(createTeamDto: CreateTeamDto) {
-    // console.log('In the service \n createTeamDto: ', createTeamDto);
-    await this.validateCreateTeam(createTeamDto);
-    // console.log('Validated');
+    const validation = await this.validateCreateTeam(createTeamDto);
+
+    if (!validation.isValid) {
+      throw new Error(validation.message);
+    }
 
     const newRole = new Team(createTeamDto);
     newRole.teamName = createTeamDto.teamName;
     newRole.companyId = createTeamDto.companyId;
     newRole.teamLeaderId = createTeamDto.teamLeaderId;
     newRole.teamMembers = createTeamDto.teamMembers;
-    console.log('New team created');
 
     return await this.teamRepository.save(newRole);
   }
@@ -133,14 +122,17 @@ export class TeamService {
     return this.teamRepository.findAll();
   }
 
-  async findById(
-    identifier: Types.ObjectId,
-  ): Promise<FlattenMaps<Team> & { _id: Types.ObjectId }> {
+  async findById(identifier: Types.ObjectId) {
     const result = await this.teamRepository.findById(identifier);
     return result;
   }
 
   async findByNameInCompany(name: string, companyId: Types.ObjectId) {
+    //checking if the company exists
+    if (!(await this.companyService.companyIdExists(companyId))) {
+      throw new Error('Company not found');
+    }
+
     const result = await this.teamRepository.findByNameInCompany(
       name,
       companyId,
@@ -156,22 +148,27 @@ export class TeamService {
     id: Types.ObjectId,
     companyId: Types.ObjectId,
   ): Promise<boolean> {
+    //checking if the company exists
+    if (!(await this.companyService.companyIdExists(companyId))) {
+      throw new Error('Company not found');
+    }
     return this.teamRepository.teamExistsInCompany(id, companyId);
   }
 
   async update(id: Types.ObjectId, updateTeamDto: UpdateTeamDto) {
-    console.log(
-      'In update service.\n id: ',
-      id,
-      '\n updateDTO: ',
-      updateTeamDto,
-    );
-    await this.validateUpdateTeam(id, updateTeamDto);
-    console.log('validate complete');
+    const validation = await this.validateUpdateTeam(id, updateTeamDto);
+
+    if (!validation.isValid) {
+      throw new Error(validation.message);
+    }
     return this.teamRepository.update(id, updateTeamDto);
   }
 
   async remove(id: Types.ObjectId): Promise<boolean> {
+    //checking if the team exists
+    if (!(await this.teamExists(id))) {
+      throw new Error('Team not found');
+    }
     return this.teamRepository.remove(id);
   }
 }
